+++ src/components/UI/BurgerMenuItem/BurgerMenuItem.js
@@ -1,0 +1,22 @@
+import PropTypes from 'prop-types';
+import React from 'react';
+
+import {toClasses} from '../../../utils';
+import styles from './BurgerMenuItem.module.scss';
+
+export const BurgerMenuItem = ({label, isActive, onClick}) => {
+  return (
+    <div
+      className={toClasses(styles.burgerMenuItem, isActive && styles.isActive)}
+      onClick={onClick}
+    >
+      {label}
+    </div>
+  );
+};
+
+BurgerMenuItem.propTypes = {
+  label: PropTypes.string,
+  isActive: PropTypes.bool,
+  onClick: PropTypes.func
+};