+++ src/config/constants.js
@@ -1,10 +1,17 @@
 export const ETHERSCAN = 'Etherscan';
 export const VOYAGER = 'Voyager';
-export const DISCORD_LINK_URL = '//discord.gg/MRjKBXtaDt';
+export const DISCORD_LINK_URL = 'https://discord.gg/MRjKBXtaDt';
 export const FETCH_TOKEN_BALANCE_MAX_RETRY = 5;
 export const STARKWARE_SITE_URL = 'https://starkware.co/';
-export const STARKGATE_DOCS_URL = 'https://starknet.io/documentation/starkgate-token-bridge';
-export const STARKNET_DOCS_URL = 'https://starknet.io/documentation/';
+export const STARKNET_SITE_URL = 'https://starknet.io';
+export const STARKNET_DOCS_URL = 'https://docs.starknet.io/docs';
+export const STARKGATE_DOCS_URL = `${STARKNET_DOCS_URL}/L1%3C%3EL2%20Communication/token-bridge`;
 export const STARKGATE_MAIL_ADDRESS = 'starkgate@starknet.io';
-export const STARKGATE_CONTRACTS_URL =
+export const STARKNET_ADDRESSES_REPO_URL = 'https://github.com/starkware-libs/starknet-addresses';
+export const STARKGATE_FRONTEND_REPO_URL = 'https://github.com/starkware-libs/starkgate-frontend';
+export const STARKGATE_CONTRACTS_REPO_URL =
   'https://github.com/starkware-libs/starkgate-contracts/tree/main/src/starkware/starknet/apps/starkgate';
+export const FEEDBACK_FORM_URL_GOERLI =
+  'https://forms.reform.app/starkware/StarkGate_Feedback/yhyalh';
+export const FEEDBACK_FORM_URL_MAINNET =
+  'https://forms.reform.app/TeRuSp/StarkGate-Feedback-Mainnet/bcoscx';