+++ src/hooks/useTransferToL2.js
@@ -1,7 +1,7 @@
 import {useCallback} from 'react';
 
 import {deposit, depositEth} from '../api/bridge';
-import {allowance, approve, balanceOf, ethBalanceOf} from '../api/erc20';
+import {allowance, approve} from '../api/erc20';
 import {ActionType, stepOf, TransferError, TransferStep, TransferToL2Steps} from '../enums';
 import {starknet} from '../libs';
 import {useDepositListener} from '../providers/EventManagerProvider';
@@ -10,8 +10,8 @@ import {useSelectedToken} from '../providers/TransferProvider';
 import {useL1Wallet, useL2Wallet} from '../providers/WalletsProvider';
 import {addToken, isEth} from '../utils';
 import {useTokenBridgeContract, useTokenContract} from './useContract';
+import {useIsMaxTotalBalanceExceeded} from './useIsMaxTotalBalanceExceeded';
 import {useLogger} from './useLogger';
-import {useMaxTotalBalance} from './useTokenConstant';
 import {useTransferToL2Tracking} from './useTracking';
 import {useTransfer} from './useTransfer';
 import {useTransferProgress} from './useTransferProgress';
@@ -22,20 +22,19 @@ export const useTransferToL2 = () => {
   const {account: l1Account, config: l1Config} = useL1Wallet();
   const {account: l2Account} = useL2Wallet();
   const {handleProgress, handleData, handleError} = useTransfer(TransferToL2Steps);
+  const {addListener, removeListener} = useDepositListener();
   const selectedToken = useSelectedToken();
   const getTokenContract = useTokenContract();
   const getTokenBridgeContract = useTokenBridgeContract();
   const progressOptions = useTransferProgress();
-  const {addListener, removeListener} = useDepositListener();
-  const maxTotalBalance = useMaxTotalBalance();
   const getL2Token = useL2Token();
+  const isMaxTotalBalanceExceeded = useIsMaxTotalBalanceExceeded();
 
   return useCallback(
     async amount => {
       const {symbol, decimals, name, tokenAddress, bridgeAddress} = selectedToken;
       const tokenContract = getTokenContract(tokenAddress);
       const bridgeContract = getTokenBridgeContract(bridgeAddress);
-      const isEthToken = isEth(symbol);
       const l2TokenAddress = getL2Token(symbol)?.tokenAddress;
 
       const readAllowance = () => {
@@ -63,7 +62,7 @@ export const useTransferToL2 = () => {
           amount,
           symbol
         });
-        const depositHandler = isEthToken ? depositEth : deposit;
+        const depositHandler = isEth(symbol) ? depositEth : deposit;
         return depositHandler({
           recipient: l2Account,
           contract: bridgeContract,
@@ -108,47 +107,45 @@ export const useTransferToL2 = () => {
         }
       };
 
-      const isMaxBalanceExceeded = async () => {
-        const tokenBridgeBalance = await (isEthToken
-          ? ethBalanceOf(bridgeAddress)
-          : balanceOf({
-              account: bridgeAddress,
-              decimals,
-              contract: tokenContract
-            }));
-        return maxTotalBalance < tokenBridgeBalance + Number(amount);
-      };
-
       try {
         logger.log('TransferToL2 called');
-        if (await isMaxBalanceExceeded()) {
-          trackReject(progressOptions.error(TransferError.MAX_TOTAL_BALANCE_ERROR));
-          logger.error(`Prevented ${symbol} deposit due to max balance exceeded`);
-          handleError(progressOptions.error(TransferError.MAX_TOTAL_BALANCE_ERROR));
-          return;
-        }
-        if (!isEthToken) {
-          logger.log('Token needs approval');
-          handleProgress(
-            progressOptions.approval(symbol, stepOf(TransferStep.APPROVE, TransferToL2Steps))
-          );
-          const allow = await readAllowance();
-          logger.log('Current allow value', {allow});
-          if (allow < amount) {
-            logger.log('Allow value is smaller then amount, sending approve tx...', {amount});
-            await sendApproval();
-          }
-        }
-        handleProgress(
-          progressOptions.waitForConfirm(
-            l1Config.name,
-            stepOf(TransferStep.CONFIRM_TX, TransferToL2Steps)
-          )
+        const {exceeded, currentTotalBalance, maxTotalBalance} = await isMaxTotalBalanceExceeded(
+          amount
         );
-        addListener(onDeposit);
-        logger.log('Calling deposit');
-        await sendDeposit();
-        await maybeAddToken(l2TokenAddress);
+        if (exceeded) {
+          const error = progressOptions.error(TransferError.MAX_TOTAL_BALANCE_ERROR, null, {
+            maxTotalBalance,
+            symbol,
+            currentTotalBalance,
+            amount
+          });
+          logger.error(`Prevented ${symbol} deposit due to maxTotalBalance exceeded`);
+          trackReject(error);
+          handleError(error);
+        } else {
+          if (!isEth(symbol)) {
+            logger.log('Token needs approval');
+            handleProgress(
+              progressOptions.approval(symbol, stepOf(TransferStep.APPROVE, TransferToL2Steps))
+            );
+            const allow = await readAllowance();
+            logger.log('Current allow value', {allow});
+            if (allow < amount) {
+              logger.log('Allow value is smaller then amount, sending approve tx...', {amount});
+              await sendApproval();
+            }
+          }
+          handleProgress(
+            progressOptions.waitForConfirm(
+              l1Config.name,
+              stepOf(TransferStep.CONFIRM_TX, TransferToL2Steps)
+            )
+          );
+          addListener(onDeposit);
+          logger.log('Calling deposit');
+          await sendDeposit();
+          await maybeAddToken(l2TokenAddress);
+        }
       } catch (ex) {
         removeListener();
         trackError(ex);
@@ -170,7 +167,7 @@ export const useTransferToL2 = () => {
       handleProgress,
       logger,
       progressOptions,
-      maxTotalBalance
+      isMaxTotalBalanceExceeded
     ]
   );
 };