+++ src/providers/AppProvider/AppProvider.js
@@ -1,5 +1,6 @@
 import PropTypes from 'prop-types';
 import React, {useReducer} from 'react';
+import {useLocation, useNavigate} from 'react-router-dom';
 
 import {useEnvs} from '../../hooks';
 import {setStorageItem} from '../../utils';
@@ -9,6 +10,13 @@ import {actions, initialState, reducer} from './app-reducer';
 export const AppProvider = ({children}) => {
   const {localStorageAcceptTermsKey} = useEnvs();
   const [state, dispatch] = useReducer(reducer, initialState);
+  const {pathname} = useLocation();
+  const navigate = useNavigate();
+  const isScrollActive = ['/terms', '/faq'].includes(pathname);
+
+  const navigateToRoute = route => {
+    pathname !== route && navigate(route);
+  };
 
   const login = () => {
     dispatch({
@@ -33,6 +41,8 @@ export const AppProvider = ({children}) => {
 
   const value = {
     ...state,
+    isScrollActive,
+    navigateToRoute,
     acceptTerms,
     login,
     logout