+++ src/providers/AppProvider/app-context.js
@@ -4,6 +4,8 @@ import {initialState} from './app-reducer';
 
 export const AppContext = createContext({
   ...initialState,
+  isScrollActive: false,
+  navigateToRoute: () => {},
   acceptTerms: () => {},
   login: () => {},
   logout: () => {}