+++ src/routes/Bridge/Bridge.js
@@ -3,10 +3,11 @@ import React, {useEffect} from 'react';
 import {setUser} from '../../analytics';
 import {Account, SelectToken, ToastProvider, Transfer} from '../../components/Features';
 import {MenuType} from '../../enums';
-import {useEnvs, useMenuTracking} from '../../hooks';
+import {useEnvs, useIsMaxTotalBalanceExceeded, useMenuTracking} from '../../hooks';
 import {BridgeProviders} from '../../providers';
 import {useMenu} from '../../providers/MenuProvider';
 import {useOnboardingModal} from '../../providers/ModalProvider';
+import {useBridgeIsFull, useSelectedToken} from '../../providers/TransferProvider';
 import {useL1Wallet, useL2Wallet} from '../../providers/WalletsProvider';
 import {getStorageItem, getMsFromHrs, setStorageItem} from '../../utils';
 import styles from './Bridge.module.scss';
@@ -18,6 +19,9 @@ export const Bridge = () => {
   const {menu, menuProps} = useMenu();
   const {account: l1account} = useL1Wallet();
   const {account: l2account} = useL2Wallet();
+  const {lockBridge, unlockBridge} = useBridgeIsFull();
+  const selectedToken = useSelectedToken();
+  const isMaxTotalBalanceExceeded = useIsMaxTotalBalanceExceeded();
 
   useEffect(() => {
     trackMenu(menu);
@@ -27,6 +31,14 @@ export const Bridge = () => {
     setUser({l1account, l2account});
     maybeShowOnboardingModal();
   }, []);
+
+  useEffect(() => {
+    async function maybeLockBridge() {
+      const {exceeded} = await isMaxTotalBalanceExceeded();
+      exceeded ? lockBridge() : unlockBridge();
+    }
+    maybeLockBridge();
+  }, [isMaxTotalBalanceExceeded, selectedToken]);
 
   const maybeShowOnboardingModal = () => {
     const onboardingTimestamp = getStorageItem(localStorageOnboardingExpirationTimestampKey);